+++ vms/avm/client.go
@@ -73,6 +73,10 @@ type Client interface {
 	//
 	// Deprecated: GetUTXOs should be used instead.
 	GetAllBalances(ctx context.Context, addr ids.ShortID, includePartial bool, options ...rpc.Option) ([]Balance, error)
+
+	// GetTxFee returns the cost to issue certain transactions
+	GetTxFee(context.Context, ...rpc.Option) (uint64, uint64, error)
+
 	// CreateAsset creates a new asset and returns its assetID
 	//
 	// Deprecated: Transactions should be issued using the
@@ -383,6 +387,12 @@ func (c *client) GetAllBalances(
 		IncludePartial: includePartial,
 	}, res, options...)
 	return res.Balances, err
+}
+
+func (c *client) GetTxFee(ctx context.Context, options ...rpc.Option) (uint64, uint64, error) {
+	res := &GetTxFeeReply{}
+	err := c.requester.SendRequest(ctx, "avm.getTxFee", struct{}{}, res, options...)
+	return uint64(res.TxFee), uint64(res.CreateAssetTxFee), err
 }
 
 // ClientHolder describes how much an address owns of an asset